refactor: Move classes into FWGameLib common
