--- Assets/Editor.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 31158fc9688340458dc109ad82488824
-timeCreated: 1723925357
