+++ Assets/FWGameLib/Common/EventSystem/IEvent.cs
@@ -1,0 +1,7 @@
+namespace FWGameLib.Common.EventSystem
+{
+    public interface IEvent
+    {
+        public void Invoke();
+    }
+}
