+++ Assets/FWGameLib/Common/EventSystem/IEvent.cs.meta
