--- Assets/_Project/DevelopmentSounds.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0d64c80983be242089385c50da063b41
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
