--- Assets/_Project/DevelopmentSounds/481775__gregorquendel__crowd-cheering-ambience-and-cheering.wav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:bebfaec9aac9fb70fe4f14a32b9a0d1ecfc967e08d959e554ae4a1332f5604ef
-size 6703244
