--- Assets/_Project/DevelopmentSounds/481775__gregorquendel__crowd-cheering-ambience-and-cheering.wav.meta
@@ -1,23 +1,0 @@
-fileFormatVersion: 2
-guid: de6e32235cded4124a2232c6e812c96b
-AudioImporter:
-  externalObjects: {}
-  serializedVersion: 7
-  defaultSettings:
-    serializedVersion: 2
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-    preloadAudioData: 0
-  platformSettingOverrides: {}
-  forceToMono: 0
-  normalize: 1
-  loadInBackground: 0
-  ambisonic: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
