+++ Assets/_Project/Scripts/Arena/ArenaBridgeLoweringCompletedEvent.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.EventSystem;
+
 public class ArenaBridgeLoweringCompletedEvent : IEvent
 {
     public ArenaBridge ArenaBridge;
