+++ Assets/_Project/Scripts/Arena/AttemptPressRedButtonEvent.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.EventSystem;
+
 public class AttemptPressRedButtonEvent : IEvent
 {
     public void Invoke()
