+++ Assets/_Project/Scripts/Arena/FallIntoVoidEvent.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.EventSystem;
+
 public class FallIntoVoidEvent : IEvent
 {
     public void Invoke()
