+++ Assets/_Project/Scripts/Arena/TutorialRedButtonPressedEvent.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.EventSystem;
+
 public class TutorialRedButtonPressedEvent : IEvent
 {
     public void Invoke()
