--- Assets/_Project/Scripts/EventSystem/IEvent.cs
@@ -1,4 +1,0 @@
-public interface IEvent
-{
-    public void Invoke();
-}
