+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/Gameplay/0-PreTutorial/PreTutorialState.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.StateMachine;
+
 public class PreTutorialState : IState
 {
     private GameManager _gameManager;
