+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/Gameplay/1-Tutorial/TutorialState.cs
@@ -1,3 +1,4 @@
+using FWGameLib.Common.StateMachine;
 using UnityEngine;
 
 public class TutorialState : IState
