+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/Gameplay/3-BossPhase1/BossAdvanceEvent.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.EventSystem;
+
 public class BossAdvanceEvent : IEvent
 {
     public void Invoke()
