+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/Gameplay/3-BossPhase1/BossPhaseState.cs
@@ -1,3 +1,4 @@
+using FWGameLib.Common.StateMachine;
 using UnityEngine;
 
 public class BossPhaseState : IState
