+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/Gameplay/GameManager.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.StateMachine;
+
 public class GameManager
 {
     public readonly StateMachine StateMachine;
