+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/Gameplay/GameOverEvent.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.EventSystem;
+
 public class GameOverEvent : IEvent
 {
     public void Invoke()
