+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/Gameplay/GameOverState.cs
@@ -1,3 +1,4 @@
+using FWGameLib.Common.StateMachine;
 using UnityEngine;
 
 public class GameOverState : IState
