+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/InGameState.cs
@@ -1,4 +1,5 @@
 using System;
+using FWGameLib.Common.StateMachine;
 using UnityEngine;
 using UnityEngine.SceneManagement;
