+++ Assets/_Project/Scripts/GameLoop/GameState/PauseState/PauseState.cs
@@ -1,3 +1,4 @@
+using FWGameLib.Common.StateMachine;
 using UnityEngine;
 using UnityEngine.SceneManagement;
