+++ Assets/_Project/Scripts/GameLoop/GameState/PauseState/QuitToMenuEvent.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.EventSystem;
+
 public class QuitToMenuEvent : IEvent
 {
     public void Invoke()
