+++ Assets/_Project/Scripts/Objects/Glass/NoteShatterEvent.cs
@@ -1,3 +1,4 @@
+using FWGameLib.Common.EventSystem;
 using UnityEngine;
 
 public class NoteShatterEvent : IEvent
