+++ Assets/_Project/Scripts/Portal/DevEnterArenaEvent.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.EventSystem;
+
 public class DevEnterArenaEvent : IEvent
 {
     public void Invoke()
