+++ Assets/_Project/Scripts/Scale/ScaleDetectionZoneEnterEvent.cs
@@ -1,3 +1,4 @@
+using FWGameLib.Common.EventSystem;
 using UnityEngine;
 
 public class ScaleDetectionZoneEnterEvent : IEvent
