+++ Assets/_Project/Scripts/Scale/TestItemScale/TestItemScale.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using FWGameLib.Common.EditorTools;
 using UnityEngine;
 
 public class TestItemScale : MonoBehaviour
