+++ Assets/_Project/Scripts/Scale/TestItemScale/TestItemScaleBoardUpdateEvent.cs
@@ -1,3 +1,5 @@
+using FWGameLib.Common.EventSystem;
+
 public class TestItemScaleBoardUpdateEvent : IEvent
 {
     public readonly int NumItems;
