--- Assets/_Project/Scripts/StateMachine.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 09c048b199804eb7a25a894b8b5a8c7f
-timeCreated: 1723925516
