--- Assets/_Project/Scripts/StateMachine/IState.cs
@@ -1,8 +1,0 @@
-// Boilerplate from: https://game.courses/bots-ai-statemachines/
-
-public interface IState
-{
-    void Tick();
-    void OnEnter();
-    void OnExit();
-}
